requirements.txt
better-profanity
bcrypt
uuid
flask
flask_cors
pymongo
pyotp
requests
pycryptodome
passlib
python-dotenv
serial